Declare variables for display and starting value
The variable that stores a reference to the display will be used quite
often so I wanted to create a variable for it so I don't have to repeat
code.
The currentTotal variable is used at the start to display a 0 when the
user loads this app or refreshes their page and later it will be used
to store totals after operators have ran.