Add listeners to buttons and logic to listeners
I sadly forgot to commit earlier, I have since added the modulous
function.
I've also added a variable which stores the last operator invoked so
operate function can come to the right conclusive action when called.
I have reorganized the variables that point to the button, operators
are now in a section away from the integer buttons.
I have added an event listener on the buttonContainer which is
event delegation and that will negate adding listeners on each of those
button variables, however this listener is composed of multiple switch
statements that are also sectioned in 2 statements, one for the
operation buttons and the others.
Right now I'm stuck in figuring out how to change the display from a
displayed total after an operation has been done and wiping it when a
number button has been pressed to just display that button and not add
it to the previous total.