Add functions for the operations of the calculator
In the return statement of each of these functions we not only return
the result of the desired operation but we convert these to numbers.
I had problems with my first attempt of this project with strings being
where I wanted integers and numbers and it collided with what I wanted
and weren't converting.