Build operate  and calculatorDisplay functions
Operate function is needed for when the user presses the '=' key to
invoke one of the operation functions and then display the result on
the display portion of the calculator.
For that to work we needed to create a function that takes a value and
displays it on the calculator, I had also created a variable in the
global scope that is referencing that element in the DOM, but now that I
look at the code I realize that variable is not needed. So that's the
next commit, omitting that variable.