Add variables for last operator and current total
So last commit I added the currentTotal variable which is important
because we need a running total in many of our operations to decipher
between new input and existing values.
We also should keep track of the last operator used... Actually I don't
think we do because the operate function uses the variable that holds
the current used operator, I'll delete it.