Add necessary styles to center calculator in body
I added styles to enlarge the buttons appropriately and applied flex to
them so that they would grow when they need to, but not shrink.
I also changed the size of the body and calculator so meet the needs
to store those buttons and to be centered vertically and horizontally.