Add functional code for sign button, remove pseudo
The code for the sign button allows the user to change the currentInput
to negative if positive and vice versa.
I have removed the pseudo code for the sign button and the back button
as they are both completed.
I have again moved the operator conditional to the bottom of this block
and I believe will work on it next.