Omit textContent from an expression being saved
This one line of code has driven me insane for the last three days, I
didn't notice it until now but am very glad I did.
This commit contains the logic for the operators being pressed, and I've
 built this same section many times over the last few days but each
broke because of the textContent following the target variable which
was giving value to the operatorPressed variable.