Change divs to buttons, style various buttons
For the focus css pseudo class to work correctly I needed to use button
elements and not divisions, so I changed the element type for every
button on the calculator.
And so I added styling to operators when clicked on they'd be
highlighted until another button is pressed and non operator buttons
will flash as they are clicked.