Change the pseudo code to reflect specifics
I thought that I would have to round decimal input, but it's only
asking us to round the output so it doesn't overflow the display.
That doesn't actually consider when we do run out of space on the left
side of the decimal so now I'm curious if they expect us to account for
that as well or not.