Change copied logic to keyboard support
I just copied the logic of the click key event press listener and
changed it to keydown events, adding arrays of possible buttons that
elicit action in the calculator.
I think this app is mostly done, however there is still a logic mistake
when the currentTotal is a zero, and when it's followed by an operator.