Add keyboard support by migrating logic
Ok I'm still learning the process for when to vs when not to commit,
and when I'm working on something and see another thing broken how to
plan that out to be the next commit and not feel pressured to do it
all in the current commit.
There are some things about this calculator I still am not content with
like how you cannot add a decimal to a total, right now it just wipes
out the total and starts a new input, I mean if the user knew how it
worked they could just add a decimal instead but that's my fault not
theirs.
Also the sign button doesn't work on the keydown events, and that's
because those keys invoke the addition and subtraction operation
functions.
Also I feel that the decimal version of the click events may interact
with the currentTotal variable when I don't want it to but have not
come across that bug when testing yet.
I'm going to call this project finished for now and move on with the
Odin project curriculum.
I'd like to state that I never looked at anyone else's calculator
code for this project, not even the example they gave us to play with,
I did however play with that example a lot to figure out logic routes.
I did have to google some basic things like how to check if a an
argument was a floating point number.
I made plenty of mistakes, first trying to use integers to be saved and
when I did switch to strings the whole range of problems the rest of my
program had because I had to migrate methods and functions to use
strings.
The biggest hiccup for me and was quite a reminder as a whole is to
check variables you're using to make sure they are what you think they
are. I was using a variable to reference the event objects however it
was actually referencing the textContent specifically and that
disallowed my logic from working, and sadly ate a couple weekend hours
until I finally came across it.
Programming is fun, don't let anyone dissuade you!